GUI/requirements.txt
pyqt6-tools
watchdog